+++ kernel/plugin/sandbox.go
@@ -17,18 +17,9 @@
 package plugin
 
 import (
-	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
-	"net/http"
-	"os"
-	"path/filepath"
 	"strconv"
-	"strings"
-	"sync"
-	"sync/atomic"
 	"time"
 
 	"github.com/dop251/goja"
@@ -36,15 +27,9 @@ import (
 	"github.com/dop251/goja_nodejs/console"
 	"github.com/dop251/goja_nodejs/require"
 	"github.com/dop251/goja_nodejs/url"
-	"github.com/gorilla/websocket"
 	"github.com/imroc/req/v3"
-	sse "github.com/r3labs/sse/v2"
 	"github.com/samber/lo"
-	"github.com/siyuan-note/filelock"
 	"github.com/siyuan-note/logging"
-	"github.com/siyuan-note/siyuan/kernel/bazaar"
-	"github.com/siyuan-note/siyuan/kernel/model"
-	"github.com/siyuan-note/siyuan/kernel/util"
 )
 
 type WebSocketState int64
@@ -80,30 +65,6 @@ func (r *CallResult) TaskResult() *TaskResult {
 		return &TaskResult{err: r.Error}
 	}
 	return &TaskResult{value: r.Value.Export()}
-}
-
-type Printer struct {
-	name string // plugin name for log prefix
-}
-
-func (p *Printer) Log(s string) {
-	go print(p.name, s, logging.LogInfof)
-}
-
-func (p *Printer) Warn(s string) {
-	go print(p.name, s, logging.LogWarnf)
-}
-
-func (p *Printer) Error(s string) {
-	go print(p.name, s, logging.LogErrorf)
-}
-
-// print logs the message line by line with the plugin name as prefix, using the provided log function.
-func print(name, s string, logf func(format string, v ...interface{})) {
-	rows := strings.Split(s, "\n")
-	for _, row := range rows {
-		logf("[plugin:%s] %s", name, row)
-	}
 }
 
 // EnableExtendModules registers extended modules (e.g. url, buffer) to the plugin's goja runtime.
@@ -149,1321 +110,6 @@ func EnableSiyuanModule(p *KernelPlugin, rt *goja.Runtime) (err error) {
 	lo.Must0(ObjectFreeze(rt, siyuan))
 
 	lo.Must0(rt.GlobalObject().Set("siyuan", siyuan))
-	return
-}
-
-// injectPlugin adds siyuan.plugin to the goja context.
-func injectPlugin(p *KernelPlugin, rt *goja.Runtime, siyuan *goja.Object) (err error) {
-	defer func() {
-		if r := recover(); r != nil {
-			err = fmt.Errorf("injectPlugin: %v", r)
-		}
-	}()
-
-	lifecycle := rt.NewObject()
-	lo.Must0(lifecycle.Set("onload", goja.Null()))
-	lo.Must0(lifecycle.Set("onloaded", goja.Null()))
-	lo.Must0(lifecycle.Set("onrunning", goja.Null()))
-	lo.Must0(lifecycle.Set("onunload", goja.Null()))
-	lo.Must0(ObjectSeal(rt, lifecycle))
-
-	plugin := rt.NewObject()
-	lo.Must0(plugin.Set("name", p.Name))
-	lo.Must0(plugin.Set("version", p.Version))
-	lo.Must0(plugin.Set("displayName", p.DisplayName))
-	lo.Must0(plugin.Set("platform", bazaar.GetCurrentBackend()))
-	lo.Must0(plugin.Set("i18n", p.I18n))
-	lo.Must0(plugin.Set("lifecycle", lifecycle))
-	lo.Must0(ObjectFreeze(rt, plugin))
-
-	lo.Must0(siyuan.Set("plugin", plugin))
-	return
-}
-
-// injectLogger adds siyuan.logger to the goja context.
-func injectLogger(p *KernelPlugin, rt *goja.Runtime, siyuan *goja.Object) (err error) {
-	defer func() {
-		if r := recover(); r != nil {
-			err = fmt.Errorf("injectLogger: %v", r)
-		}
-	}()
-
-	logger := rt.NewObject()
-
-	lo.Must0(logger.Set("trace", rt.ToValue(loggerWrapper(p, logging.LogTracef))))
-	lo.Must0(logger.Set("debug", rt.ToValue(loggerWrapper(p, logging.LogDebugf))))
-	lo.Must0(logger.Set("info", rt.ToValue(loggerWrapper(p, logging.LogInfof))))
-	lo.Must0(logger.Set("warn", rt.ToValue(loggerWrapper(p, logging.LogWarnf))))
-	lo.Must0(logger.Set("error", rt.ToValue(loggerWrapper(p, logging.LogErrorf))))
-
-	lo.Must0(ObjectFreeze(rt, logger))
-
-	lo.Must0(siyuan.Set("logger", logger))
-	return
-}
-
-// injectEvent adds siyuan.event to the goja context.
-func injectEvent(p *KernelPlugin, rt *goja.Runtime, siyuan *goja.Object) (err error) {
-	defer func() {
-		if r := recover(); r != nil {
-			err = fmt.Errorf("injectEvent: %v", r)
-		}
-	}()
-
-	event := rt.NewObject()
-
-	lo.Must0(event.Set("handler", goja.Null()))
-
-	lo.Must0(event.Set("emit", rt.ToValue(func(call goja.FunctionCall, rt *goja.Runtime) goja.Value {
-		promise, resolve, reject := rt.NewPromise()
-
-		runErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-			if len(call.Arguments) < 2 {
-				err = fmt.Errorf("topic and event required")
-				return
-			}
-
-			topic := call.Argument(0).String()
-			if topic == "" {
-				err = fmt.Errorf("topic required")
-				return
-			}
-
-			event := call.Argument(1).Export()
-
-			p.bus.Publish(topic, event)
-			return
-		}, func(rt *goja.Runtime, result any, err error) {
-			if lo.IsNil(err) {
-				if resolveErr := resolve(result); resolveErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.event.emit resolve: %v", p.Name, resolveErr)
-				}
-			} else {
-				if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.event.emit reject: %v", p.Name, rejectErr)
-				}
-			}
-		})
-		if runErr != nil {
-			logging.LogErrorf("[plugin:%s] siyuan.event.emit worker run: %v", p.Name, runErr)
-		}
-
-		return rt.ToValue(promise)
-	})))
-
-	lo.Must0(ObjectSeal(rt, event))
-
-	lo.Must0(siyuan.Set("event", event))
-	return
-}
-
-// injectStorage adds siyuan.storage.* methods for scoped file CRUD.
-func injectStorage(p *KernelPlugin, rt *goja.Runtime, siyuan *goja.Object) (err error) {
-	defer func() {
-		if r := recover(); r != nil {
-			err = fmt.Errorf("injectStorage: %v", r)
-		}
-	}()
-
-	baseDir := filepath.Join(util.DataDir, "storage", "petal", p.Name)
-
-	resolvePath := func(relPath string) (abs string, err error) {
-		abs = filepath.Join(baseDir, filepath.Clean(relPath))
-		if !(abs == baseDir || strings.HasPrefix(abs, baseDir+string(filepath.Separator))) {
-			err = fmt.Errorf("siyuan.storage: path traversal not allowed")
-		}
-		return
-	}
-
-	storage := rt.NewObject()
-
-	// siyuan.storage.get(path) -> Promise<{text, json, arrayBuffer}>
-	lo.Must0(storage.Set("get", rt.ToValue(func(call goja.FunctionCall, rt *goja.Runtime) goja.Value {
-		promise, resolve, reject := rt.NewPromise()
-
-		runErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-			var path string
-			if len(call.Arguments) >= 1 && goja.IsString(call.Argument(0)) {
-				path = call.Argument(0).String()
-			} else {
-				err = fmt.Errorf("path required")
-				return
-			}
-
-			abs, resolveErr := resolvePath(path)
-			if resolveErr != nil {
-				err = resolveErr
-				return
-			}
-
-			go func() (result []byte, err error) {
-				defer func() {
-					if r := recover(); r != nil {
-						err = fmt.Errorf("panic during siyuan.storage.get: %v", r)
-					}
-					p.worker.Run(func(rt *goja.Runtime) (_ any, _ error) {
-						if err != nil {
-							return nil, err
-						}
-
-						content := rt.NewObject()
-						if err = ObjectSetDataMethods(p, rt, content, result); err != nil {
-							return nil, err
-						}
-
-						return content, nil
-					}, func(rt *goja.Runtime, result any, err error) {
-						if lo.IsNil(err) {
-							if resolveErr := resolve(result); resolveErr != nil {
-								logging.LogErrorf("[plugin:%s] siyuan.storage.get resolve: %v", p.Name, resolveErr)
-							}
-						} else {
-							if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-								logging.LogErrorf("[plugin:%s] siyuan.storage.get reject: %v", p.Name, rejectErr)
-							}
-						}
-					})
-				}()
-
-				data, readErr := filelock.ReadFile(abs)
-				if readErr != nil {
-					err = readErr
-					return
-				}
-
-				result = data
-				return
-			}()
-
-			return
-		}, func(rt *goja.Runtime, result any, err error) {
-			if !lo.IsNil(err) {
-				if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.storage.get reject: %v", p.Name, rejectErr)
-				}
-			}
-		})
-		if runErr != nil {
-			logging.LogErrorf("[plugin:%s] siyuan.storage.get worker run: %v", p.Name, runErr)
-		}
-
-		return rt.ToValue(promise)
-	})))
-
-	// siyuan.storage.put(path, content) -> Promise<void>
-	lo.Must0(storage.Set("put", rt.ToValue(func(call goja.FunctionCall, rt *goja.Runtime) goja.Value {
-		promise, resolve, reject := rt.NewPromise()
-
-		runErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-			if len(call.Arguments) < 2 {
-				err = fmt.Errorf("path and content required")
-				return
-			}
-			path := call.Argument(0).String()
-			content := call.Argument(1).String()
-			abs, resolveErr := resolvePath(path)
-			if resolveErr != nil {
-				err = resolveErr
-				return
-			}
-
-			go func() (result any, err error) {
-				defer func() {
-					if r := recover(); r != nil {
-						err = fmt.Errorf("panic during siyuan.storage.put: %v", r)
-					}
-
-					p.worker.Run(func(rt *goja.Runtime) (_ any, _ error) {
-						if lo.IsNil(err) {
-							if resolveErr := resolve(result); resolveErr != nil {
-								logging.LogErrorf("[plugin:%s] siyuan.storage.put resolve: %v", p.Name, resolveErr)
-							}
-						} else {
-							if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-								logging.LogErrorf("[plugin:%s] siyuan.storage.put reject: %v", p.Name, rejectErr)
-							}
-						}
-						return
-					}, nil)
-				}()
-
-				if mkdirErr := os.MkdirAll(filepath.Dir(abs), 0755); mkdirErr != nil {
-					err = fmt.Errorf("failed to make directory: %w", mkdirErr)
-					return
-				}
-				if writeErr := filelock.WriteFile(abs, []byte(content)); writeErr != nil {
-					err = fmt.Errorf("failed to write file: %w", writeErr)
-					return
-				}
-				return
-			}()
-
-			return
-		}, func(rt *goja.Runtime, result any, err error) {
-			if !lo.IsNil(err) {
-				if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.storage.put reject: %v", p.Name, rejectErr)
-				}
-			}
-		})
-		if runErr != nil {
-			logging.LogErrorf("[plugin:%s] siyuan.storage.put worker run: %v", p.Name, runErr)
-		}
-
-		return rt.ToValue(promise)
-	})))
-
-	// siyuan.storage.remove(path) -> Promise<void>
-	lo.Must0(storage.Set("remove", rt.ToValue(func(call goja.FunctionCall, rt *goja.Runtime) goja.Value {
-		promise, resolve, reject := rt.NewPromise()
-
-		runErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-			if len(call.Arguments) < 1 {
-				err = fmt.Errorf("path required")
-				return
-			}
-			path := call.Argument(0).String()
-			abs, resolveErr := resolvePath(path)
-			if resolveErr != nil {
-				err = resolveErr
-				return
-			}
-			if abs == baseDir {
-				err = fmt.Errorf("cannot remove storage root")
-				return
-			}
-
-			go func() (result any, err error) {
-				defer func() {
-					if r := recover(); r != nil {
-						err = fmt.Errorf("panic during siyuan.storage.remove: %v", r)
-					}
-
-					p.worker.Run(func(rt *goja.Runtime) (_ any, _ error) {
-						if lo.IsNil(err) {
-							if resolveErr := resolve(result); resolveErr != nil {
-								logging.LogErrorf("[plugin:%s] siyuan.storage.remove resolve: %v", p.Name, resolveErr)
-							}
-						} else {
-							if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-								logging.LogErrorf("[plugin:%s] siyuan.storage.remove reject: %v", p.Name, rejectErr)
-							}
-						}
-						return
-					}, nil)
-				}()
-
-				if removeErr := os.RemoveAll(abs); removeErr != nil {
-					err = fmt.Errorf("failed to remove: %w", removeErr)
-					return
-				}
-				return
-			}()
-
-			return
-		}, func(rt *goja.Runtime, result any, err error) {
-			if !lo.IsNil(err) {
-				if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.storage.remove reject: %v", p.Name, rejectErr)
-				}
-			}
-		})
-		if runErr != nil {
-			logging.LogErrorf("[plugin:%s] siyuan.storage.remove worker run: %v", p.Name, runErr)
-		}
-
-		return rt.ToValue(promise)
-	})))
-
-	// siyuan.storage.list(path) -> Promise<Entry[]>
-	lo.Must0(storage.Set("list", rt.ToValue(func(call goja.FunctionCall, rt *goja.Runtime) goja.Value {
-		promise, resolve, reject := rt.NewPromise()
-
-		runErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-			if len(call.Arguments) < 1 {
-				err = fmt.Errorf("path required")
-				return
-			}
-			path := call.Argument(0).String()
-			abs, resolveErr := resolvePath(path)
-			if resolveErr != nil {
-				err = resolveErr
-				return
-			}
-
-			go func() (result any, err error) {
-				defer func() {
-					if r := recover(); r != nil {
-						err = fmt.Errorf("panic during siyuan.storage.list: %v", r)
-					}
-					p.worker.Run(func(rt *goja.Runtime) (_ any, _ error) {
-						if lo.IsNil(err) {
-							if resolveErr := resolve(result); resolveErr != nil {
-								logging.LogErrorf("[plugin:%s] siyuan.storage.list resolve: %v", p.Name, resolveErr)
-							}
-						} else {
-							if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-								logging.LogErrorf("[plugin:%s] siyuan.storage.list reject: %v", p.Name, rejectErr)
-							}
-						}
-						return
-					}, nil)
-				}()
-
-				entries, readErr := os.ReadDir(abs)
-				if readErr != nil {
-					err = fmt.Errorf("failed to read directory: %w", readErr)
-					return
-				}
-
-				results := make([]R, 0, len(entries))
-				for _, entry := range entries {
-					info, infoErr := entry.Info()
-					if infoErr != nil {
-						continue
-					}
-					results = append(results, R{
-						"name":      entry.Name(),
-						"isDir":     info.IsDir(),
-						"isSymlink": util.IsSymlink(entry),
-						"updated":   info.ModTime().Unix(),
-					})
-				}
-
-				result = results
-				return
-			}()
-
-			return
-		}, func(rt *goja.Runtime, result any, err error) {
-			if !lo.IsNil(err) {
-				if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.storage.list reject: %v", p.Name, rejectErr)
-				}
-			}
-		})
-		if runErr != nil {
-			logging.LogErrorf("[plugin:%s] siyuan.storage.list worker run: %v", p.Name, runErr)
-		}
-
-		return rt.ToValue(promise)
-	})))
-
-	lo.Must0(ObjectFreeze(rt, storage))
-
-	lo.Must0(siyuan.Set("storage", storage))
-	return
-}
-
-// injectRpc adds siyuan.rpc method for RPC method registration.
-func injectRpc(p *KernelPlugin, rt *goja.Runtime, siyuan *goja.Object) (err error) {
-	defer func() {
-		if r := recover(); r != nil {
-			err = fmt.Errorf("injectRpc: %v", r)
-		}
-	}()
-
-	rpc := rt.NewObject()
-
-	lo.Must0(rpc.Set("bind", rt.ToValue(func(call goja.FunctionCall, rt *goja.Runtime) goja.Value {
-		promise, resolve, reject := rt.NewPromise()
-
-		runErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-			var name string
-			var method goja.Callable
-			var descriptions []string
-
-			if len(call.Arguments) < 2 {
-				err = fmt.Errorf("method name and function required")
-				return
-			}
-			nameArg := call.Argument(0)
-			methodArg := call.Argument(1)
-			descArgs := call.Arguments[2:]
-
-			if goja.IsString(nameArg) {
-				name = nameArg.String()
-			} else {
-				err = fmt.Errorf("first argument must be method name string")
-				return
-			}
-
-			if methodJs, ok := goja.AssertFunction(methodArg); ok {
-				method = methodJs
-			} else {
-				err = fmt.Errorf("second argument must be a function")
-			}
-
-			descriptions = make([]string, len(descArgs))
-			for i, a := range descArgs {
-				descriptions[i] = a.String()
-			}
-
-			err = p.bindRpcMethod(name, method, descriptions...)
-			return
-		}, func(rt *goja.Runtime, result any, err error) {
-			if lo.IsNil(err) {
-				if resolveErr := resolve(result); resolveErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.rpc.bind resolve: %v", p.Name, resolveErr)
-				}
-			} else {
-				if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.rpc.bind reject: %v", p.Name, rejectErr)
-				}
-			}
-		})
-		if runErr != nil {
-			logging.LogErrorf("[plugin:%s] siyuan.rpc.bind worker run: %v", p.Name, runErr)
-		}
-
-		return rt.ToValue(promise)
-	})))
-
-	lo.Must0(rpc.Set("unbind", rt.ToValue(func(call goja.FunctionCall, rt *goja.Runtime) goja.Value {
-		promise, resolve, reject := rt.NewPromise()
-
-		runErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-			var name string
-			if len(call.Arguments) < 1 {
-				err = fmt.Errorf("method name required")
-				return
-			}
-
-			nameArg := call.Argument(0)
-			if goja.IsString(nameArg) {
-				name = nameArg.String()
-			} else {
-				err = fmt.Errorf("first argument must be method name string")
-				return
-			}
-
-			err = p.unbindRpcMethod(name)
-			return
-		}, func(rt *goja.Runtime, result any, err error) {
-			if lo.IsNil(err) {
-				if resolveErr := resolve(result); resolveErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.rpc.unbind resolve: %v", p.Name, resolveErr)
-				}
-			} else {
-				if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.rpc.unbind reject: %v", p.Name, rejectErr)
-				}
-			}
-		})
-		if runErr != nil {
-			logging.LogErrorf("[plugin:%s] siyuan.rpc.unbind worker run: %v", p.Name, runErr)
-		}
-
-		return rt.ToValue(promise)
-	})))
-
-	lo.Must0(rpc.Set("broadcast", rt.ToValue(func(call goja.FunctionCall, rt *goja.Runtime) goja.Value {
-		promise, resolve, reject := rt.NewPromise()
-
-		runErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-			if len(call.Arguments) < 1 {
-				err = fmt.Errorf("method required")
-				return
-			}
-
-			var method string
-			if m := call.Argument(0); goja.IsString(m) {
-				method = m.String()
-			}
-
-			var params util.Optional[any]
-			arg := call.Argument(1)
-			if goja.IsUndefined(arg) {
-				params.Value = nil
-				params.Exists = false
-				params.IsNull = false
-			} else if goja.IsNull(arg) {
-				params.Value = nil
-				params.Exists = true
-				params.IsNull = true
-			} else {
-				params.Value = arg.Export()
-				params.Exists = true
-				params.IsNull = false
-			}
-
-			p.BroadcastNotification(method, params)
-			return
-		}, func(rt *goja.Runtime, result any, err error) {
-			if lo.IsNil(err) {
-				if resolveErr := resolve(result); resolveErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.rpc.broadcast resolve: %v", p.Name, resolveErr)
-				}
-			} else {
-				if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.rpc.broadcast reject: %v", p.Name, rejectErr)
-				}
-			}
-		})
-		if runErr != nil {
-			logging.LogErrorf("[plugin:%s] siyuan.rpc.broadcast worker run: %v", p.Name, runErr)
-		}
-
-		return rt.ToValue(promise)
-	})))
-
-	lo.Must0(ObjectFreeze(rt, rpc))
-
-	lo.Must0(siyuan.Set("rpc", rpc))
-	return
-}
-
-// injectClient adds siyuan.server to the goja context.
-func injectClient(p *KernelPlugin, rt *goja.Runtime, siyuan *goja.Object) (err error) {
-	defer func() {
-		if r := recover(); r != nil {
-			err = fmt.Errorf("injectClient: %v", r)
-		}
-	}()
-
-	client := rt.NewObject()
-
-	lo.Must0(client.Set("fetch", rt.ToValue(func(call goja.FunctionCall, rt *goja.Runtime) goja.Value {
-		promise, resolve, reject := rt.NewPromise()
-
-		runErr := p.worker.Run(func(rt *goja.Runtime) (_ any, err error) {
-			var path string
-			if len(call.Arguments) > 0 && goja.IsString(call.Argument(0)) {
-				path = call.Argument(0).String()
-			} else {
-				err = fmt.Errorf("path required")
-				return
-			}
-
-			if !strings.HasPrefix(path, "/") {
-				err = fmt.Errorf("path must start with /")
-				return
-			}
-
-			method := "GET"
-			headers := map[string]string{}
-			var bodyString string
-			var bodyBytes []byte
-			if len(call.Arguments) > 1 {
-				init := call.Argument(1)
-				if init != nil && !goja.IsUndefined(init) && !goja.IsNull(init) {
-					if initObj := init.ToObject(rt); initObj != nil {
-						if m := initObj.Get("method"); m != nil && goja.IsString(m) {
-							method = m.String()
-						}
-
-						if h := initObj.Get("headers"); h != nil && !goja.IsUndefined(h) && !goja.IsNull(h) {
-							if hObj := h.ToObject(rt); hObj != nil {
-								if hMap, ok := h.Export().(map[string]string); ok {
-									for k, v := range hMap {
-										headers[k] = v
-									}
-								}
-							}
-						}
-
-						if b := initObj.Get("body"); b != nil && !goja.IsUndefined(b) && !goja.IsNull(b) {
-
-							if goja.IsString(b) {
-								bodyString = b.String()
-							} else {
-								body := b.Export()
-								if arrayBuffer, ok := body.(goja.ArrayBuffer); ok {
-									bodyBytes = arrayBuffer.Bytes()
-								}
-							}
-						}
-					}
-				}
-			}
-
-			go func() (err error) {
-				defer func() {
-					if r := recover(); r != nil {
-						err = fmt.Errorf("panic during siyuan.client.fetch: %v", r)
-					}
-
-					if err != nil {
-						p.worker.Run(func(rt *goja.Runtime) (_ any, _ error) {
-							if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-								logging.LogErrorf("[plugin:%s] siyuan.client.fetch reject: %v", p.Name, rejectErr)
-							}
-							return
-						}, nil)
-					}
-				}()
-
-				targetURL := fmt.Sprintf("http://127.0.0.1:%s%s", util.ServerPort, path)
-				r := httpClient.R()
-				for k, v := range headers {
-					r.SetHeader(k, v)
-				}
-				r.SetHeader(model.XAuthTokenKey, p.token)
-
-				if bodyString != "" {
-					r.SetBody(bodyString)
-				} else if len(bodyBytes) > 0 {
-					r.SetBody(bodyBytes)
-				}
-
-				resp, sendErr := r.Send(method, targetURL)
-				if sendErr != nil {
-					err = sendErr
-					return
-				}
-
-				defer resp.Body.Close()
-				body, readErr := io.ReadAll(resp.Body)
-				if readErr != nil {
-					err = fmt.Errorf("failed to read response body: %w", readErr)
-					return
-				}
-
-				responseHeader := map[string]string{}
-				for k, vs := range resp.Header {
-					responseHeader[k] = strings.Join(vs, ", ")
-				}
-
-				runErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-					response := rt.NewObject()
-					lo.Must0(response.Set("url", rt.ToValue(path)))
-					lo.Must0(response.Set("ok", rt.ToValue(resp.StatusCode >= 200 && resp.StatusCode < 300)))
-					lo.Must0(response.Set("status", rt.ToValue(resp.StatusCode)))
-					lo.Must0(response.Set("statusText", rt.ToValue(resp.Status)))
-					lo.Must0(response.Set("headers", rt.ToValue(responseHeader)))
-					lo.Must0(ObjectSetDataMethods(p, rt, response, body))
-					result = response
-					return
-				}, func(rt *goja.Runtime, result any, err error) {
-					if lo.IsNil(err) {
-						if resolveErr := resolve(result); resolveErr != nil {
-							logging.LogErrorf("[plugin:%s] siyuan.client.fetch resolve: %v", p.Name, resolveErr)
-						}
-					} else {
-						if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-							logging.LogErrorf("[plugin:%s] siyuan.client.fetch reject: %v", p.Name, rejectErr)
-						}
-					}
-				})
-				if runErr != nil {
-					err = runErr
-					return
-				}
-
-				return
-			}()
-
-			return
-		}, func(rt *goja.Runtime, _ any, err error) {
-			if !lo.IsNil(err) {
-				if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.client.fetch reject: %v", p.Name, rejectErr)
-				}
-			}
-		})
-		if runErr != nil {
-			logging.LogErrorf("[plugin:%s] siyuan.client.fetch worker run: %v", p.Name, runErr)
-		}
-
-		return rt.ToValue(promise)
-	})))
-
-	lo.Must0(client.Set("socket", rt.ToValue(func(call goja.FunctionCall, rt *goja.Runtime) goja.Value {
-		promise, resolve, reject := rt.NewPromise()
-
-		runErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-			var path string
-			if len(call.Arguments) > 0 && goja.IsString(call.Argument(0)) {
-				path = call.Argument(0).String()
-			} else {
-				err = fmt.Errorf("path required")
-				return
-			}
-
-			if !strings.HasPrefix(path, "/") {
-				err = fmt.Errorf("path must start with /")
-				return
-			}
-
-			var protocols []string
-			if len(call.Arguments) > 1 {
-				proto := call.Argument(1)
-				if !goja.IsNull(proto) && !goja.IsUndefined(proto) {
-					if protoObj := proto.ToObject(rt); protoObj != nil && protoObj.ClassName() == "Array" {
-						if arr, ok := proto.Export().([]interface{}); ok {
-							for _, v := range arr {
-								protocols = append(protocols, fmt.Sprintf("%v", v))
-							}
-						}
-					} else {
-						protocols = []string{proto.String()}
-					}
-				}
-			}
-
-			var conn atomic.Pointer[websocket.Conn]
-			var readyState atomic.Int64
-			var bufferedAmount atomic.Int64
-			messagesCh := make(chan WebSocketMessage, 256)
-			connReadyCh := make(chan *websocket.Conn, 1)
-			senderDoneCh := make(chan struct{})
-
-			wsURL := fmt.Sprintf("ws://127.0.0.1:%s%s", util.ServerPort, path)
-			wsHeader := http.Header{}
-			wsHeader.Set(model.XAuthTokenKey, p.token)
-			if len(protocols) > 0 {
-				wsHeader.Set("Sec-WebSocket-Protocol", strings.Join(protocols, ", "))
-			}
-
-			wsObj := rt.NewObject()
-
-			lo.Must0(wsObj.Set("binaryType", rt.ToValue("arraybuffer")))
-			lo.Must0(wsObj.Set("bufferedAmount", rt.ToValue(0)))
-			lo.Must0(wsObj.Set("extensions", rt.ToValue("")))
-			lo.Must0(wsObj.Set("protocol", rt.ToValue("")))
-			lo.Must0(wsObj.Set("readyState", rt.ToValue(int64(WebSocketReadyStateConnecting))))
-			lo.Must0(wsObj.Set("url", rt.ToValue("wsURL")))
-
-			lo.Must0(wsObj.Set("onopen", goja.Null()))
-			lo.Must0(wsObj.Set("onping", goja.Null()))
-			lo.Must0(wsObj.Set("onpong", goja.Null()))
-			lo.Must0(wsObj.Set("onerror", goja.Null()))
-			lo.Must0(wsObj.Set("onmessage", goja.Null()))
-			lo.Must0(wsObj.Set("onclose", goja.Null()))
-
-			invokeWsHook := func(name string, args ...goja.Value) {
-				hook := wsObj.Get(name)
-				if fn, ok := goja.AssertFunction(hook); ok {
-					if _, callErr := fn(wsObj, args...); callErr != nil {
-						logging.LogErrorf("[plugin:%s] ws hook %q: %v", p.Name, name, callErr)
-					}
-				}
-			}
-
-			setReadyState := func(rt *goja.Runtime, state WebSocketState) {
-				readyState.Store(int64(state))
-				wsObj.Set("readyState", rt.ToValue(state))
-			}
-
-			updateBufferedAmount := func(rt *goja.Runtime, delta int) {
-				bufferedAmount.Add(int64(delta))
-				wsObj.Set("bufferedAmount", rt.ToValue(bufferedAmount.Load()))
-			}
-
-			setReadyState(rt, WebSocketReadyStateConnecting)
-
-			lo.Must0(wsObj.Set("send", rt.ToValue(func(sendCall goja.FunctionCall, rt *goja.Runtime) goja.Value {
-				sendPromise, sendResolve, sendReject := rt.NewPromise()
-
-				sendRunErr := p.worker.Run(func(rt *goja.Runtime) (_ any, err error) {
-					var messageData []byte
-					var messageType int
-					if len(sendCall.Arguments) >= 1 {
-						data := sendCall.Argument(0)
-						if arrayBuffer, ok := data.Export().(goja.ArrayBuffer); ok {
-							messageType = websocket.BinaryMessage
-							messageData = arrayBuffer.Bytes()
-						} else {
-							messageType = websocket.TextMessage
-							messageData = []byte(data.String())
-						}
-					}
-
-					state := WebSocketState(readyState.Load())
-					if state == WebSocketReadyStateClosing || state == WebSocketReadyStateClosed {
-						err = fmt.Errorf("WebSocket is not open (state: %d)", state)
-						return
-					}
-
-					updateBufferedAmount(rt, len(messageData))
-
-					done := make(chan FunctionResult[int], 1)
-
-					select {
-					case messagesCh <- WebSocketMessage{
-						t:    messageType,
-						d:    messageData,
-						done: done,
-					}:
-						go func() {
-							select {
-							case result := <-done:
-								p.worker.Run(func(rt *goja.Runtime) (_ any, err error) {
-									if result.Error == nil {
-										updateBufferedAmount(rt, -result.Value)
-									} else {
-										err = result.Error
-									}
-									return
-								}, func(rt *goja.Runtime, result any, err error) {
-									if lo.IsNil(err) {
-										if resolveErr := sendResolve(result); resolveErr != nil {
-											logging.LogErrorf("[plugin:%s] siyuan.client.socket.send resolve: %v", p.Name, resolveErr)
-										}
-									} else {
-										if rejectErr := sendReject(rt.NewGoError(err)); rejectErr != nil {
-											logging.LogErrorf("[plugin:%s] siyuan.client.socket.send reject: %v", p.Name, rejectErr)
-										}
-									}
-								})
-							case <-senderDoneCh:
-							}
-						}()
-					default:
-						updateBufferedAmount(rt, -len(messageData))
-						err = fmt.Errorf("WebSocket send buffer full")
-					}
-					return
-				}, func(rt *goja.Runtime, _ any, err error) {
-					if lo.IsNil(err) {
-					} else {
-						if rejectErr := sendReject(rt.NewGoError(err)); rejectErr != nil {
-							logging.LogErrorf("[plugin:%s] siyuan.client.socket.send reject: %v", p.Name, rejectErr)
-						}
-					}
-				})
-				if sendRunErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.client.socket.send worker run: %v", p.Name, sendRunErr)
-				}
-
-				return rt.ToValue(sendPromise)
-			})))
-
-			lo.Must0(wsObj.Set("ping", rt.ToValue(func(pingCall goja.FunctionCall, rt *goja.Runtime) goja.Value {
-				pingPromise, pingResolve, pingReject := rt.NewPromise()
-
-				pingRunErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-					var pingData string
-					if len(pingCall.Arguments) > 0 && !goja.IsUndefined(pingCall.Argument(0)) {
-						pingData = pingCall.Argument(0).String()
-					}
-					if c := conn.Load(); c != nil {
-						_, err = p.writeWebSocketMessage(c, websocket.PingMessage, []byte(pingData))
-					}
-					return
-				}, func(rt *goja.Runtime, result any, err error) {
-					if lo.IsNil(err) {
-						if resolveErr := pingResolve(result); resolveErr != nil {
-							logging.LogErrorf("[plugin:%s] siyuan.client.socket.ping resolve: %v", p.Name, resolveErr)
-						}
-					} else {
-						if rejectErr := pingReject(rt.NewGoError(err)); rejectErr != nil {
-							logging.LogErrorf("[plugin:%s] siyuan.client.socket.ping reject: %v", p.Name, rejectErr)
-						}
-					}
-				})
-				if pingRunErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.client.socket.ping worker run: %v", p.Name, pingRunErr)
-				}
-
-				return rt.ToValue(pingPromise)
-			})))
-
-			lo.Must0(wsObj.Set("pong", rt.ToValue(func(pongCall goja.FunctionCall, rt *goja.Runtime) goja.Value {
-				pongPromise, pongResolve, pongReject := rt.NewPromise()
-
-				pongRunErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-					var pongData string
-					if len(pongCall.Arguments) > 0 && !goja.IsUndefined(pongCall.Argument(0)) {
-						pongData = pongCall.Argument(0).String()
-					}
-					if c := conn.Load(); c != nil {
-						_, err = p.writeWebSocketMessage(c, websocket.PongMessage, []byte(pongData))
-					}
-					return
-				}, func(rt *goja.Runtime, result any, err error) {
-					if lo.IsNil(err) {
-						if resolveErr := pongResolve(result); resolveErr != nil {
-							logging.LogErrorf("[plugin:%s] siyuan.client.socket.pong resolve: %v", p.Name, resolveErr)
-						}
-					} else {
-						if rejectErr := pongReject(rt.NewGoError(err)); rejectErr != nil {
-							logging.LogErrorf("[plugin:%s] siyuan.client.socket.pong reject: %v", p.Name, rejectErr)
-						}
-					}
-				})
-				if pongRunErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.client.socket.pong worker run: %v", p.Name, pongRunErr)
-				}
-
-				return rt.ToValue(pongPromise)
-			})))
-
-			lo.Must0(wsObj.Set("close", rt.ToValue(func(closeCall goja.FunctionCall, rt *goja.Runtime) goja.Value {
-				closePromise, closeResolve, closeReject := rt.NewPromise()
-
-				closeRunErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-					code := websocket.CloseNormalClosure
-					var reason string
-					if len(closeCall.Arguments) > 0 && !goja.IsUndefined(closeCall.Argument(0)) {
-						code = int(closeCall.Argument(0).ToInteger())
-					}
-					if len(closeCall.Arguments) > 1 && !goja.IsUndefined(closeCall.Argument(1)) {
-						reason = closeCall.Argument(1).String()
-					}
-					if c := conn.Load(); c != nil {
-						_, err = p.writeWebSocketMessage(c, websocket.CloseMessage, websocket.FormatCloseMessage(code, reason))
-					}
-					return
-				}, func(rt *goja.Runtime, result any, err error) {
-					if lo.IsNil(err) {
-						if resolveErr := closeResolve(result); resolveErr != nil {
-							logging.LogErrorf("[plugin:%s] siyuan.client.socket.close resolve: %v", p.Name, resolveErr)
-						}
-					} else {
-						if rejectErr := closeReject(rt.NewGoError(err)); rejectErr != nil {
-							logging.LogErrorf("[plugin:%s] siyuan.client.socket.close reject: %v", p.Name, rejectErr)
-						}
-					}
-				})
-				if closeRunErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.client.socket.close worker run: %v", p.Name, closeRunErr)
-				}
-
-				return rt.ToValue(closePromise)
-			})))
-
-			// Start a goroutine to send messages from the channel to the WebSocket connection
-			go func() {
-				var c *websocket.Conn
-				select {
-				case c = <-connReadyCh:
-				case <-senderDoneCh:
-					return
-				}
-				for {
-					select {
-					case m := <-messagesCh:
-						amount, writeErr := p.writeWebSocketMessage(c, m.t, m.d)
-						m.done <- FunctionResult[int]{Value: amount, Error: writeErr}
-					case <-senderDoneCh:
-						return
-					}
-				}
-			}()
-
-			go func() (err error) {
-				defer func() {
-					close(senderDoneCh)
-					if r := recover(); r != nil {
-						err = fmt.Errorf("panic during siyuan.client.socket: %v", r)
-					}
-
-					if err != nil {
-						p.worker.Run(func(rt *goja.Runtime) (_ any, _ error) {
-							event := rt.NewObject()
-							event.Set("type", rt.ToValue("error"))
-							event.Set("error", rt.NewGoError(err))
-							invokeWsHook("onerror", event)
-							return
-						}, nil)
-					}
-				}()
-
-				dialer := websocket.Dialer{}
-				c, _, dialErr := dialer.Dial(wsURL, wsHeader)
-				if dialErr != nil {
-					err = dialErr
-					return
-				}
-
-				defer func() {
-					p.UntrackSocket(c)
-					c.Close()
-				}()
-				p.TrackSocket(c, false)
-
-				c.SetPingHandler(func(data string) error {
-					_, runError := p.worker.RunSync(func(rt *goja.Runtime) (result any, err error) {
-						event := rt.NewObject()
-						event.Set("type", rt.ToValue("ping"))
-						event.Set("data", rt.ToValue(data))
-						invokeWsHook("onping", event)
-						return
-					})
-					return runError
-				})
-				c.SetPongHandler(func(data string) error {
-					_, runError := p.worker.RunSync(func(rt *goja.Runtime) (result any, err error) {
-						event := rt.NewObject()
-						event.Set("type", rt.ToValue("pong"))
-						event.Set("data", rt.ToValue(data))
-						invokeWsHook("onpong", event)
-						return
-					})
-					return runError
-				})
-				c.SetCloseHandler(func(code int, reason string) error {
-					_, runError := p.worker.RunSync(func(rt *goja.Runtime) (result any, err error) {
-						setReadyState(rt, WebSocketReadyStateClosing)
-						event := rt.NewObject()
-						event.Set("type", rt.ToValue("close"))
-						event.Set("code", rt.ToValue(int64(code)))
-						event.Set("reason", rt.ToValue(reason))
-						invokeWsHook("onclose", event)
-						return
-					})
-					return runError
-				})
-
-				connReadyCh <- c
-
-				_, runErr := p.worker.RunSync(func(rt *goja.Runtime) (result any, err error) {
-					conn.Store(c)
-					wsObj.Set("protocol", c.Subprotocol())
-					setReadyState(rt, WebSocketReadyStateOpen)
-
-					// Emit the open event
-					event := rt.NewObject()
-					event.Set("type", rt.ToValue("open"))
-					invokeWsHook("onopen", event)
-					return
-				})
-				if runErr != nil {
-					err = runErr
-					return
-				}
-
-				for {
-					messageType, data, readErr := c.ReadMessage()
-					if readErr != nil {
-						if websocket.IsUnexpectedCloseError(readErr, websocket.CloseNormalClosure, websocket.CloseGoingAway) {
-							err = readErr
-						}
-						return
-					}
-					switch messageType {
-					case websocket.TextMessage:
-						runErr := p.worker.Run(func(rt *goja.Runtime) (_ any, _ error) {
-							event := rt.NewObject()
-							event.Set("type", rt.ToValue("text"))
-							event.Set("data", rt.ToValue(string(data)))
-							invokeWsHook("onmessage", event)
-							return
-						}, nil)
-						if runErr != nil {
-							err = runErr
-							return
-						}
-					case websocket.BinaryMessage:
-						runErr := p.worker.Run(func(rt *goja.Runtime) (_ any, _ error) {
-							event := rt.NewObject()
-							event.Set("type", rt.ToValue("binary"))
-							event.Set("data", rt.ToValue(rt.NewArrayBuffer(data)))
-							invokeWsHook("onmessage", event)
-							return
-						}, nil)
-						if runErr != nil {
-							err = runErr
-							return
-						}
-					}
-				}
-			}()
-
-			result = wsObj
-			return
-		}, func(rt *goja.Runtime, result any, err error) {
-			if lo.IsNil(err) {
-				if resolveErr := resolve(result); resolveErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.client.socket resolve: %v", p.Name, resolveErr)
-				}
-			} else {
-				if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.client.socket reject: %v", p.Name, rejectErr)
-				}
-			}
-		})
-		if runErr != nil {
-			logging.LogErrorf("[plugin:%s] siyuan.client.socket worker run: %v", p.Name, runErr)
-		}
-
-		return rt.ToValue(promise)
-	})))
-
-	lo.Must0(client.Set("event", rt.ToValue(func(call goja.FunctionCall, rt *goja.Runtime) goja.Value {
-		promise, resolve, reject := rt.NewPromise()
-
-		runErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-			var path string
-			if len(call.Arguments) > 0 && goja.IsString(call.Argument(0)) {
-				path = call.Argument(0).String()
-			} else {
-				err = fmt.Errorf("path required")
-				return
-			}
-
-			if !strings.HasPrefix(path, "/") {
-				err = fmt.Errorf("path must start with /")
-				return
-			}
-
-			var readyState atomic.Int64
-			esURL := fmt.Sprintf("http://127.0.0.1:%s%s", util.ServerPort, path)
-
-			esObj := rt.NewObject()
-
-			setReadyState := func(state EventSourceState) {
-				readyState.Store(int64(state))
-				esObj.Set("readyState", rt.ToValue(state))
-			}
-
-			invokeEsHook := func(name string, args ...goja.Value) {
-				hook := esObj.Get(name)
-				if fn, ok := goja.AssertFunction(hook); ok {
-					if _, callErr := fn(esObj, args...); callErr != nil {
-						logging.LogErrorf("[plugin:%s] es hook %q: %v", p.Name, name, callErr)
-					}
-				}
-			}
-
-			setReadyState(EventSourceConnecting)
-			lo.Must0(esObj.Set("url", rt.ToValue(path)))
-			lo.Must0(esObj.Set("onopen", goja.Null()))
-			lo.Must0(esObj.Set("onmessage", goja.Null()))
-			lo.Must0(esObj.Set("onclose", goja.Null()))
-			lo.Must0(esObj.Set("onerror", goja.Null()))
-
-			ctx, cancel := context.WithCancel(context.Background())
-			sseID := p.TrackSSE(cancel)
-
-			var closeOnce sync.Once
-			close := func() {
-				closeOnce.Do(func() {
-					cancel()
-					p.UntrackSSE(sseID)
-				})
-			}
-
-			lo.Must0(esObj.Set("close", rt.ToValue(func(goja.FunctionCall) goja.Value {
-				setReadyState(EventSourceClosed)
-				close()
-				return goja.Undefined()
-			})))
-
-			go func() (err error) {
-				defer func() {
-					if r := recover(); r != nil {
-						err = fmt.Errorf("panic during siyuan.client.event: %v", r)
-					}
-
-					close()
-
-					p.worker.Run(func(rt *goja.Runtime) (_ any, _ error) {
-						if EventSourceState(readyState.Load()) != EventSourceClosed {
-							if err != nil && !errors.Is(err, context.Canceled) {
-								event := rt.NewObject()
-								event.Set("type", rt.ToValue("error"))
-								event.Set("error", rt.NewGoError(err))
-								invokeEsHook("onerror", event)
-							}
-							setReadyState(EventSourceClosed)
-						}
-						return
-					}, nil)
-				}()
-
-				sseClient := sse.NewClient(esURL)
-				sseClient.Headers[model.XAuthTokenKey] = p.token
-
-				sseClient.OnConnect(func(_ *sse.Client) {
-					p.worker.Run(func(rt *goja.Runtime) (_ any, _ error) {
-						setReadyState(EventSourceOpen)
-						event := rt.NewObject()
-						event.Set("type", rt.ToValue("open"))
-						invokeEsHook("onopen", event)
-						return
-					}, nil)
-				})
-
-				sseClient.OnDisconnect(func(_ *sse.Client) {
-					p.worker.Run(func(rt *goja.Runtime) (_ any, _ error) {
-						setReadyState(EventSourceClosed)
-						event := rt.NewObject()
-						event.Set("type", rt.ToValue("close"))
-						invokeEsHook("onclose", event)
-						return
-					}, nil)
-				})
-
-				err = sseClient.SubscribeRawWithContext(ctx, func(msg *sse.Event) {
-					p.worker.Run(func(rt *goja.Runtime) (_ any, _ error) {
-						typ := "message"
-						if len(msg.Event) > 0 {
-							typ = string(msg.Event)
-						}
-						event := rt.NewObject()
-						event.Set("type", rt.ToValue(typ))
-						event.Set("data", rt.ToValue(string(msg.Data)))
-						event.Set("lastEventId", rt.ToValue(string(msg.ID)))
-						invokeEsHook("onmessage", event)
-						return
-					}, nil)
-				})
-				return
-			}()
-
-			result = esObj
-			return
-		}, func(rt *goja.Runtime, result any, err error) {
-			if lo.IsNil(err) {
-				if resolveErr := resolve(result); resolveErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.client.event resolve: %v", p.Name, resolveErr)
-				}
-			} else {
-				if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.client.event reject: %v", p.Name, rejectErr)
-				}
-			}
-		})
-		if runErr != nil {
-			logging.LogErrorf("[plugin:%s] siyuan.client.event worker run: %v", p.Name, runErr)
-		}
-
-		return rt.ToValue(promise)
-	})))
-
-	lo.Must0(ObjectFreeze(rt, client))
-
-	lo.Must0(siyuan.Set("client", client))
-	return
-}
-
-// injectServer adds siyuan.server to the goja context.
-func injectServer(p *KernelPlugin, rt *goja.Runtime, siyuan *goja.Object) (err error) {
-	defer func() {
-		if r := recover(); r != nil {
-			err = fmt.Errorf("injectServer: %v", r)
-		}
-	}()
-
-	http := rt.NewObject()
-	ws := rt.NewObject()
-	es := rt.NewObject()
-
-	lo.Must0(http.Set("handler", goja.Null()))
-	lo.Must0(ws.Set("handler", goja.Null()))
-	lo.Must0(es.Set("handler", goja.Null()))
-
-	lo.Must0(ObjectSeal(rt, http))
-	lo.Must0(ObjectSeal(rt, ws))
-	lo.Must0(ObjectSeal(rt, es))
-
-	private := rt.NewObject()
-
-	lo.Must0(private.Set(string(RequestTypeHTTP), http))
-	lo.Must0(private.Set(string(RequestTypeWS), ws))
-	lo.Must0(private.Set(string(RequestTypeSSE), es))
-
-	lo.Must0(ObjectFreeze(rt, private))
-
-	server := rt.NewObject()
-
-	lo.Must0(server.Set("private", private))
-
-	lo.Must0(ObjectFreeze(rt, server))
-
-	lo.Must0(siyuan.Set("server", server))
 	return
 }
 
@@ -1616,63 +262,6 @@ func NewDataObject(p *KernelPlugin, rt *goja.Runtime, data []byte) (*goja.Object
 		return nil, err
 	}
 	return obj, nil
-}
-
-// loggerWrapper returns a JS-callable function that logs a message at the given level.
-// Logging is synchronous (in-process, no I/O), so resolve is called inline without worker.Run.
-func loggerWrapper(p *KernelPlugin, logf func(format string, args ...any)) func(goja.FunctionCall, *goja.Runtime) goja.Value {
-	return func(call goja.FunctionCall, rt *goja.Runtime) goja.Value {
-		promise, resolve, reject := rt.NewPromise()
-
-		runErr := p.worker.Run(func(rt *goja.Runtime) (result any, err error) {
-			parts := make([]string, 0, len(call.Arguments))
-			for _, arg := range call.Arguments {
-				if goja.IsString(arg) {
-					parts = append(parts, arg.String())
-					continue
-				}
-
-				if arg == nil {
-					parts = append(parts, "null")
-					continue
-				}
-
-				if goja.IsUndefined(arg) || goja.IsNull(arg) {
-					parts = append(parts, arg.String())
-					continue
-				}
-
-				argObj := arg.ToObject(rt)
-				if argObj != nil {
-					if argJson, marshalErr := argObj.MarshalJSON(); marshalErr == nil {
-						parts = append(parts, string(argJson))
-						continue
-					}
-				}
-
-				parts = append(parts, arg.String())
-			}
-			msg := strings.Join(parts, " ")
-
-			go print(p.Name, msg, logf)
-			return
-		}, func(rt *goja.Runtime, result any, err error) {
-			if lo.IsNil(err) {
-				if resolveErr := resolve(result); resolveErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.logger resolve: %v", p.Name, resolveErr)
-				}
-			} else {
-				if rejectErr := reject(rt.NewGoError(err)); rejectErr != nil {
-					logging.LogErrorf("[plugin:%s] siyuan.logger reject: %v", p.Name, rejectErr)
-				}
-			}
-		})
-		if runErr != nil {
-			logging.LogErrorf("[plugin:%s] siyuan.logger worker run: %v", p.Name, runErr)
-		}
-
-		return rt.ToValue(promise)
-	}
 }
 
 // getJsContextValue safely retrieves a nested value from the plugin's JS context, returning nil if any step fails.
